New: Possibility to record actions with Cucapp
Cucapp now provides an API to record a set of events and play it again.
To use it, the user need to call the function load_cucapp_record(path) and then call the function start_record(path).
Once he has done, the users needs to call the function stop_record() and then save_record(fileName).
This is going to save a JSON file with the informations of the CPEvents made and other informations (identifier, clicked offset) about the clicked control.

Then, you can play your record with the function play_record(file_name, cucumber_path). This is going to reproduce what the user just did, in the same time.

To save the events, a category of CPWindow and CPApplication is used. The category is only loaded when you call the function load_cucapp_record(path).