New: more informations when clicking on a view with the CLI
Previously, when clicking a view we only displayed the informations of the first view caught.
Now we log the informations of the views caught for this click. It means we will display the first view, its parent, its parent and so on.
We will only display the view which have a cucappIdentifier, text, label, tag, title, objectValue, placeholderString or identifier.