New: added the command cucapp
This PR adds the command cucapp to the cappuccino environment.
To install cucapp, you now need to do jake install. Jake will install the necessary gem (Cucumber, Thin, Nokogiri, Launchy, JSON) for cucapp. Then Jake will copy the library and the binary of cucapp in narwhal (most of the time it will be in /usr/local/narwhal/packages or in your capp_env). Once it's installed, you have now the command cucapp.

This command allows you to create an testing architecture in your cappuccino application. It will create the folder features which contains the cucapp library, the file hooks.rb for your setUp and tearDown, the file env.rb for your global env of your test, the file encumber_category.rb and Cucumber+Extensions.j for having the possibility to create new interaction with cappuccino and a exemple of a cucumber feature (a test).

The Jakefile of your application will change when you do Cucapp as well. It will contain the jake task cucumber-test (surprise it launches the cucumber test). When you launch this command, jake will create a soft link of Cucapp (from usr/local/narwhal/packages/cucapp for example), use this soft link to launch the tests and then remove this link when the tests are done.