New: Added a set of default steps
This commit adds a set of steps. These steps can be used to write functional or unit tests.
It will help the user to not rewrite steps each time he wants to create tests for an application.
These steps provides an api to easily click (right, left or double), scroll, drag and drop, hit a key.
This PR comes with a cappuccino_mappings file, this file is used to map the cappuccino control name to a human name (for example field for CPTextField). It gives a more human readability to the tests itself.